requirements.txt
deta==1.1.0
fastapi==0.94.1
python-dotenv==1.0.0
